Fixes for activity tracking (Blocking Stack + use in ActivityComputer)
Removed TraceActivies from TraceLog.cs.   (will move TraceActivty to ActivityComputer next)
Added BlockingStack() to CSwtiches so at the time a CSwitch blocks you can get the stack.

Added BlockingStack to the 'Any Stacks' view.
Added HasBlockedStack to Events view.

Used this to make ActivityComputer better.   Now when we are in Parked in the Threadpool we will implicitly create a new activity.
This fixes a number of issues.