Fix problem where Debug build fails in VS2015 (but not Debug VS2013, and does not fail in Release builds either).
It would fail with about 30 missing symbol (mostly exception goo, but also allocation and other primitive services).   It does not fail on release

Tracked it down to the fact that I explicitly declare the DllMain entry point.  This avoids the library's DLL main, which probably avoided some references which woudl have defined the undefined things (arguably a bug but whatever).     I only had to remove the DLLMain for Debug, so that is what I did.  It does not seem to affect functionality.