Fix empty stacks bug, remove line number tracking from FastStream
There was a bug where if the sample did not have a stack trace, the parser
would move onto the next non-empty line and start using that as its frame. So sometimes
a sample header would become a frame, and, unsurprisingly, the module and symbol would not
be found, resulting in an empty module and symbol, and a display of just '!'. I also noticed
that keeping track of line numbers in the FastStream was useless since I can always
check timestamps on samples to know where the parser is messing up