Fix bugs parsing various unusual kinds of strings (found in the WinINet provider).
As part of this I did a rather non-trivial reoganization/simplification of the counted string parsing logic for
registered providers.   It is now much more regular and thus more likely to work properly in all cases.

I still need to do a bit more testing on it, but I wanted to 'get it out there' as it does seem to work well enough
already (and is an improvement over what we have..