Fix issue where opening a .trace.zip file failes with a file in use error when opening CPU stacks.
The issue was twofold.

First we were opening ZIP files in exclusive mode when reading, that is bad, use ZipFile.Open which does the right thing (allows other simultaneous readers)
However a related issue is that we did not close the ZIP file when the CTF open failed.   This left the file open so that the next open (to open stacks) would fail.