As I explored more complex uses of NeedProcesses and NeedLoadedDotNetRuntimes I found that the check for when to add callbacks was too aggressive.  I built a NeedThreadPool* extension for CAP and was receiving mulitiple JIT callbacks because for a given source callbacks were added mulitple times.  For more complex scenarios NeedLoadedDotNetRuntimes may be called multiple times witht the same source.
Running with a live source trace session for a few hours hits repeat Process PIDs.  The current moved TraceProcess code did not have all the right details to make that work correctly.
In addition the Debug.Asserts in TraceProcessesExentions were not working as designed and the one that really matters (in NeedProcesses) is still in place and guards what is needed most.

Significant perf improvement.  There is special code in PerfInfoSample that was running for EVERY TRACE when it is only needed with configured for MutableStacks.  This change guards that code behind a check which SIGNIFICANTLY (reducing the processing time of a 1.5 gig file to ~1 minute from multiple hours [frankly I did not let it finish]).