Catch exceptions in TraceCodeAddresses.LookupSymbolsForModule.
Normally it is not good for APIs to catch all exceptions because you lose error information.
However for symbol lookup we have LOTS of logging in the model and pretty much EVERY
user does want it to be robust (if it fails, you can continue, and users want that).  So it does
makes sense to make it the default.