Fix for issue #166 Exception when opening CPU View
The problem is that trace had a process whose name began with a number.
Fixed it to not be so senstive (however if the process has a name that has a space and then a number it it will still be confused).
Still this is much better, and arguably good enough