Unbreak debug build and tests
This change also disables the .NET Core build because it is broken on debug and we want to fix it using .NET 2.0 features.
We will turn it on separated (it is not stricktly needed)

There was a change that added a BuildTime field and needed the baselines to be updates.  There was also a Assert in MemInfo because nameing was not consistent.

I also removed GetTypeInfo() (it is a no-op put in for .NET core support), call because it is no longer necessary for .NET 2.0 and was causing grief.