Make GeneralParsing test less OS dependent
In the GeneralParsingTest we do use the OS to look up how to decode events.
Because of this if you run on different versions of the OS you will get differnet results which cause the test to fail.

Fixing this by simply ignoring events that are likley to be affected by this (MSNT_SystemTrace events).
This is not perfect, but there is a tradeoff here, we DO want to test these OS lookup paths to, so we only want to exclude events that are likely to be 'new' and thus might vary from version to version.