Simplify the logic associated binary Support Files (#210)
* Simplify the logic associated binary Support Files

There are two nuget packages that we used to carry binary files that are needed
by perfview but are not built by this repo.    There was some copy logic in those
that was causing grief for Visual Studio.   Simplify this to avoid the need for the
copy.

Upgraded these support nuget packages to version 1.0.2.   This is mostly a no-op (the binaries did not change, the the possible exception of small changes to OSExtentions.dll)
but I needed to update these nuget packages to simplify them.

Also unbroke the logic for generating these support files.

Finally removed some .SLN files that have not been used in a long time and are probably confusing.

* Better docs on the process of updating the support Files packages.