Fix problem opening DiagSession files when VS 2015 was not installed
A dll DiagnosticsHub.Packaging.Interop.dll was not deployed in the correct place (it was placed in the architecture directory)
This caused the logic in that package not to find the architecture specific DLL and it fails with a 'argument can't be null' exception.

Fix this.  Had to upgrade the PerfView.SupportFIles package to 1.0.3 to do so.  Updated these DLLs to the newest version of
the DiagnosticHub Dlls while I was at it.