Several minor updates
Added a link to more documenation about collecting on Linux machines.
Updated logic so that /OnlyProviders will collect addtion events if the user asks for stacks on those events (there was logic but it was geared to old syntax).
Fix a bug with creating large data that might exceed 4GB (happened with some large heap dumps).
Fixed bug where /CpuCounters argument was lost if PerfView had to elevate.