Fix Incorrect GCStats information when reading an ETL file directly
Turns out that some ETL files have and incorrect sessionEnd time
(sometimes it is wildly too small).   This caused GCStats to give
wildly incorrect results.

The incorrect results were mostly because we forced all timestamps
to be less than sessionEndTime.   We now stop doing this.

Note that this does not affect ETLX (perfView) traces because when
the ETLX file is created we 'fix; the sessionEndTime if it is incurrect
(smaller than any of the event timestamps).