Fix native symbol lookup failure
When we updated the build scripts to use the new 2.0 CSPROJ format
we lost the fact that we emed the Interop code for the Interop.Dia2lib.dll
This caused any native symbol lookups to fail (since it is needed to interface to msdia140.dll)

It is not clear that the new build system supports embedding Interop code but we
did not really need it (We coudl just load the interop DLL for real).   However for this
to work the DLL needs to be named Dia2Lib.    It is unclear why this DLL was named
Interop.Dia2lib.dll before so we are just going to change it.

Do do this rename, we need to update the Support library package (since Interop.Dia2lib.dll
is in there), and then update to refrence the new version (with the rename).