Remove dependency on System.Diagnostics.Eventing.EventDescriptor
We had been using System.Diagnostics.Eventing.EventDescriptor in some of our PINVOKE structures that happend to exist in the Desktop Runtime
However this type was removed for NetStandard.

There was no particular reason to take this depencency (it just happened to already exist).  This change changes its use to a local defintion
(using the name EVENT_DESRIPTOR (which is how it is defined in C++ and using the C++ names for the fields (basically what we would have
done if it was just another structure needed for PINOVOKE).

This change is in peparation for compiling TraceEvent against NetStandard (not just desktop).

Also so fixed a mintor unreleated test issue that showed up during testing (added a missing [Fact] attribute).