Copy native PerfViewSupport dlls to bin directory
This fixes issue #356.   Note that the problem was only when running under the debugger.
Becasue the native DLLs were not copied to the bin directory, they were not found when
a diagSession file was opened, leading to the null argument exception.    Outside the debugger
the DLLs are unpacked by PerfVIew itself and that works properly.