Fix test break
We have some useful asserts that check consistancy of a TraceEventParser but they
do assume that certain conventions are followed that the IIS provider is not following
(That the name of the C# event is the same as the event name).    This is not strictly
correct, but we can keep the checking working by normalizing event names before
doing the check.   This unbreaks the code.

Also fixed a HTML warning.