Fixes for SysCall.
Basically in the SysCall events don't have process or thread IDs but we DO collect stacks, which have these things.
We already had logic that uses these stacks to fix events without thread IDs and so we just extend this for Process IDs.

Also changed it so that the stack is on the 'SysCall' exit since that has better information (the return code).