Moves FetchSymbolsForProcess to be a user command.
This was put in as a 'direct' command understood by PerfView directly.   Moved it to be
a user command.  We have wanted to do this for a while because this command is not really
that central, and we do it now becuase it is convinient for making a .NET Core version
of the PerfView collector (for Nano-OS).

FetchSymbolsForProcess has dependencies on code that are related to the GUI which is
not available on .NET Core.   By moving to be an extension, we make the 'core' parts
of perfview less entagled with the GUI which makes the port easier.

I also happened to rename PerfViewEventSource.cs to PerfViewLogger.cs, because that
is the name of hte class in that file.   This is just cleanup.