This represents the first attempt at building PerfViewCollect
Basically what we want is a command line version of PerfVIew that runs on .NET core, and thus on Windows-Nano.

This is the first shot at it (it builds, but does not run properly).   This change adds the new PerfVIewCollect project,
and #ifdefs some code with either !DOTNET_CORE (when PerfView uses things that are on the Desktop runtime
but not on .NET Core (mostly PerfCounters)), or !PERFVIEW_COLLECT (when there is a depedency on some GUI
things (typcially messsagebox to warn the user).

This is the first iteration.  I woudl like to trim the number of #ifdefs down with better factoring, but that is a later step.