Fixes to make basic ETW collection work.
At this point it look like it could be a useful program.   (I have not tried it on Nano however).

The biggest issue right now seems to be deployment things.  It works
if you deploy all the DLLs, however ideally we get the auto-deploy part working.
Currently it looks like .NET Core does not support the AssemblyResolve event.
We should be able to work around it...