Change AddDispatch Hook to allow wrapping hooks
Jeff checked in AddDispatch but after the review I wanted the API to easily support 'wrapping' (e.g. lock) scenarios.
(That is the hook is called which can set somethign up, call the original dispatch, and then tear it down).
I talked with Jeff, and I he is OK with this change.