Make TraceLog real time sessions handle exceptions well.
Today if there is an exception when processing some events in a TraceLog real time session. it wil tear down the process.

This is because we need another thread to process kernel events and if an exception
happens there it reaches the base of that thread.    Instead of using Threads, use
Tasks which automatically capture exceptions thrown in them and re-raise them
when the Task is waited upon.  This is what we want here, so the fix was very easy.