Allow Events to have too more fields than have been serialized. (#555)
Also updated the docs a bit.

Previously it was an error if an event had more fileds than were serialized.
Now it simply makes the 'extra' fields empty.   This allows versioning (you can add more fileds, but if
some older EventSource does not create the new fields, it is not an error, you simply just don't get the
information.  This makes for better versioning.