Fixed the ManifestData Event Pretty Printing
The ManifestData event prints an ETW manifest when it is pretty printed.
The problem is that this manifest does not represent the data in the event
but the BEST manifest that we have accuulated to that point.  This is misleading.

This is not easy to fix because the manifest is actually spread over multiple events.
Yet it still is very nice to have the manifest as part of the dump of that event.

The compromise I did here, is to put a warning in as well as the timestamp of
the event that WAS the defining event.