Some small PerfView improvements
Added NoNGenPdbs (useful for .NET Core scenarios which don't need them)
Fixed issues with logging the CPU Counter IDs, (eg Cache Misses ...) (needs more work but this is a start).
Allowed ProcessorNumber to be viewed on any event.
Decoded another variation of the kernel interrupt Service routine event (which is used on newer OSes).
Fixed bug where collection log is not put in the ZIP if there are no symbol files (like when you use NoNGenPdbs)