Fixes an issue where sometimes native DLLs can't get their symbols resolved.
As part of merging ETL files certain Image* events are added to add among other things
the GUID that uniquely identifies the DLL and its cooresponding PDB.
Previous to this change, we required that these Image* events be place BEFORE the
ImageLoad event they are for.

Turns out that is not always the case.   Fix it so that either order is fine.