Add DLR support for TraceEvent dynamic properties
The Dynamic Language Runtime approach won't be as efficient as the
TraceParserGen approach, but it is convenient for prototyping or
accessing event data from PowerShell.

(see the updates to the documentation for a bit more info)

The DLR is supported in .NET Standard 2.0, but not 1.6, so we #if it out
for 1.6.

Addresses #617.