Fixed issue where PDBS for DLL were not found.
The issue was that a previous change to attribute the PDB
signature to the DLL fixed one case but broke another.
In particualr if several DLL loads have the same timestamp
the logic can get confused.    Fixed this case.

Also added the number of usec for the QPC tick count when
you dump and event (some events use these ticks).