It seems we were passing tests becaues Asserts were being swallowed. Fix this.
Basically we were turning asserts to throws, which then got caught and swallowed.
Fix it so that we fail the test oin that case.

Also fix these tests to be more uniform about how we unpack the data files we comare to.   improve the test so that it is more helpful when there are failures.

Finally there were problems that were being ignored that needed to be fixed.
In particular, we were assigning two stacks to the same event (becase there were events with the same QPC time).   Fixed this by disambiguating based on processor number (this is the change in TraceLog.cs).