Added support for Goto Source for NetPerf files
Most of the logic to support Goto Source 'just works' for NetPerf files but there was an early-out check that the file was an ETL file.    Fix this.

Added a small improvement to what we point in PerfView's log when capturing a GC heap dump.

Added a small improvement to avoid manipulating a kernel session when KernelEvents == None.

Added 'the 'Type' field when showing ClrThreadSamples in the 'Any Stacks' view.