Fix bug in reading PerfView.xml files
There is an issue where doubles are written in one local and read in another.  This was already fixed in the data, but there was a lingering problem in the FilterState (e.g. the values in the Start and End TextBoxes are also persisted in the XML).   This fix only makes the reader robust (that is it will ignore bad data.