Fix dev-time load failure loading KernelTraceControl.
OSExtentions needs to be 'next' to the architecture directories where KernelTraceControl.dll lives for it to load
them properly.   This works for published .NET Core apps, but by default 'dotnet run' runs nuget DLLs out of
the nuget cache (it does not copy them to the app directory).   This breaks the loading of KernelTraceControl.dll.

We already manually copy the native files and some interop files to the output directory in the TraceEvent nuget
package.   Here we simply add OSExtentions.dll to that list so it is loaded from the same place even in the dev-time
scenario.