Support IncompleteAsyncMethod (at least a bit).
.NET Core added an event that helps diagnose some kinds of async deadlock
(where code that should have triggered a task completion simply drops the
task that should have been signaled).     See release notes for more.

TraceEvent synthesizes a Version number to determine if one EventSource
manifest is 'better' than another.   Change it so that if you add a new event
the version goes a a lot, which allows you to also delete some entries and still
have a better version (see comments for more).