Use the OS support for TraceLogging ETW parsing
There is old code in PerfView dating to the time when the self-describing (TraceLogging) format was only supported on Windows 10, so TraceEvent parsed that logic itself.    However this logic has been back ported to even Win7 and Win8 so there is no good reason to not use the built-in OS support.

This change removes the hand parsing, and simply uses the OS support.