Fix bug in a complex Self-describing ETW example
There was an example where a class with 4 nullable fields was not
being parsed correctly.    This was caused by a misunderstanding
about how to parse propertyInfo information (InType is not valid
if the property has struct type, see change RegisteredTraceEventParser.cs)

Along the way I also spotted a bug where the cache for property fetches could be fooled by nested structs.   Added a 'useCache' argument so
the cache is only used for top level structs.