Added the /FocusProcess qualifier.
This allows you to limit collection of non-kernel events (and rundown) to just one process
(specified either by name (exe file name + extention) or processID.

In addtion
Moved from deprecated licenseurl tag in the Nuget package.
Code to avoid null references in on case with activities.

Added a popup if there are time ording problems in the ETL file.