Allow building the MemoryGraph and Utilities dlls.
In the past PerfView code was used to build libraries that were used in Visual Studio.   The
DLLs that were built were Microsoft.Diagnostics.Utilties and Microsoft.Diagnostics.MemoryGraph.

These DLLs are not built as part of perfView (PerfView incorperates the soruce in the the PerfView.exe directly).  The build logic for these two DLLs is still in the repo, but it bit-rotted.
I have updated the MemoryGraph.csproj and Utiltiies.csproj to restore the ability to build these
DLL as the VS MAY find this useful (and it was easy to to do),    This may not turn out to be
useful, but at least this way the CSPROJ files are not completely out of date.

Note that all the changes in *.cs file are simply to removing the unneeded 'internal' specififer that causes a compile error if UTILITIES_PUBLIC is turned on.