Code changes to enable use of Segmented Lists
These changes permit us to use Segmented Lists in place of Growable Arrays to reduce the chance of running out of memory when using the MemoryGraph class. In order to not break the existing API of Perfview I made GrowableArray and SegmentedList share a common interface. When you create a Graph or MemoryGraph, you can use the overloaded constructor to specify that it should store the internal data in a SegmentedList instead of the GrowableArray. This way I do not require any client to change its current programs that use PerfView. The other classes I create are used to support the use of streams with segmented lists because they run into the same problem.

We added the sln and csproj files to HeapDump to enable the building of
the dll. Since it is only consumed by VS it only targets net45, since
net standard is missing some of its dependencies.