Non-semantic refactoring changes in Graph.cs
These changes should not change behavior at all.
This is in preparation for better sampling (sampling when
graph is created rather than afterward.

Also added code for a CompressedStream (ifdefed out) so we don't lose it (it should be useful for *.netperf files and
I don't want to lose the work).