Small fixes
Fix logic that allows EventSources to skip the * syntax (e.g. /ProviderName=XXX will assume XXX is
and eventSource (with a warning in the log), if it could not find it in the OS table).    This was the
intent of code added long ago but it was broken.

Renamed Insure->Ensure

Small fix in inlining processing logic.