Fix bug where PerfView fails to launch.
We added code to give windows owners (so that windows end up on a related screen when using multiple screens)
However it turns out you can't give a window an owner if the owner was never display, and we do precisely this
when we pop a EULA dialog on startup.  This causes a silent crash (because the normal windows to display errors
also fail).    Fix this by catching errors and proceeding without setting owners.

Also updated to use a newer SupportFiles (this just adds a PDB), and update the PerfView version number.