Changes window ownership for better z-ordering
Previous #758 fixed window creation so that child windows would be on the same monitor as their parents. This change has the side effect of making it impossible to put parent windows on top of child windows.

This change instead puts the window on the screen where the mouse cursor is at the time the window is created. This isn't perfect; if you double click something and quickly move the mouse you can wind up with the window on the "wrong" monitor. I think this is a better compromise than the current situation.

Fixes #865