Eliminating unnecessary buffering
When streaming the EventPipe data any attempt to buffer beyond the minimum required data to read the next EventBlock can cause event processing to stall. There is no guarantee about when or if more bytes will ever be written to the stream. FastSerialization was doing this with an automatic 16KB look-ahead that caused several iterations of EventCounter data to accumulate in the buffer before any of it was made available in the event callbacks. I modified StreamReaderWriter so that it does not buffer beyond the caller's minimum requested data (modulo 8byte alignment) when dealing with non-seekable streams.