Enable EventPipe streaming (#915)
* Enable EventPipe streaming

1) The IO underlaying Deserializer assumes that the stream has bounded length and is seekable (like a file) but neither of these is true for a network stream. I made a few modifications that avoid those assumptions, but the result feels fragile - it depends on the input stream not having references that require seeking to occur. This appears true for the netperf file format but not true for arbitrary graphs serialized with FastSerialization.

2) EventPipeEventSource needed a new constructor that accepted a Stream rather than a file

3) Dynamic parsing for EventPipe events appeared to be broken, but its possible in my relative new-ness to TraceEvent work I just don't understand how all the pieces are supposed to fit together.

a) From what I can tell EventPipeEventSource had an atypical member _eventParser and it had a private contract to send it all the metadata packets for parsing. However a parser won't register new templates with the source if nobody has subscribed to those event types and it appears impossible to subscribe to the private parser. At the same time DynamicTraceEventParser had a 2nd EventPipeParser but that one would never be asked to resolve an unhandled template because ExternalTraceEventParser from which it derives prevents more than one parser of the same type from registering.

From my read Parsers that want to get metadata from the stream have to extract it from the sequence of TraceEvent objects that are produced by the source, however EventPipe metadata packets aren't surfaced as TraceEvent objects. Although its plausible that they could be abstracted that way, it would be a breaking change to produce a different set of events than we did in the past and direct analysis of the metadata packets probably isn't that interesting. This meant EventPipeParser needed an alternate way of accessing metadata so I added a side-band communication where EventPipeParser knows how to access templates that have been pre-populated on EventPipeSource.

Most of the new code EventPipeSource is a direct move from EventPipeParser. Specifically code in ReadEventParametersAndBuildTemplate, ParseFields, and GetOpcodeFromEventName is unchanged.

b) The dynamic parser wasn't transitively applying its subscriptions to the child EventPipeParser, but that was easily solved by following the same pattern as the registeredParser that was already there.

* Added streaming unit tests and misc fixes

1) Updated the code in StreamReaderWriter so that the seek and non-seek case are a little more symmetrical.
2) Added the new streaming EventPipe unit tests which promptly started hitting new issues that needed fixing below...
3) Updated the naming format of the ProcessName field on EventPipeEventSource so that it matches the convention used in TraceLog
4) Avoided an assert that was firing on TraceEvent objects that initialized taskName to "" instead of null. This case occurs when
    the parser encounters native runtime events which do not provide an event name in the EventPipe metadata.
5) Fixed an assert that was triggering when trying to listed to the "All" event on ClrTraceEventParser. This parser has multiple
    templates that match every field except the name which violates asserts in the Dispatcher subscribe code that expect each new
    template to be unique relative all previously added ones.

* Eliminating unnecessary buffering

When streaming the EventPipe data any attempt to buffer beyond the minimum required data to read the next EventBlock can cause event processing to stall. There is no guarantee about when or if more bytes will ever be written to the stream. FastSerialization was doing this with an automatic 16KB look-ahead that caused several iterations of EventCounter data to accumulate in the buffer before any of it was made available in the event callbacks. I modified StreamReaderWriter so that it does not buffer beyond the caller's minimum requested data (modulo 8byte alignment) when dealing with non-seekable streams.