Fix GC suspension parser to coexist with stack sampler
This fixes issue #924. The parser expects to see only a single threaded sequence of SuspendStart, SuspendStop, RestartStart, RestartStop but using GC and stack sampling simultaneously can interleave that pattern on two different threads. This confuses the parser and breaks the GC stats.

The issue is fixed by filtering out all SuspendStart events that don't have a GC reason and then filtering out further suspension events if they are not on the same thread as the SuspendStart event.