Upgrade NetPerf format (step 2)
To allow streaming large files we need to be able to run FastSerialization without interning the objects that are being read or memory will leak over time. However FastSerialization required every object to have a type and if you tried to emit them using BeginPrivateObject to avoid interning the code didn't correctly recognize that a NullReference should be interpreted as typeof(SerializationType). I am assuming the omission was accidental and is now fixed.