Fix parsing of ETLX files larger than 2GB
StreamLabels serialized in the FastSerialization format are 4 byte unsigned integers, but the FastSerialization library reads them using ReadInt32() so values above 2GB appear negative. Prior to PR #952 StreamLabel was a uint so (StreamLabel)ReadInt32() restored the correct unsigned semantics to the value. In PR #952 StreamLabel became a long which meant values above 2GB remained negative. Attempting to Goto() a negative StreamLabel put the reader in a bad state which lead to the observed error later in Fill().

The fix casts the int first to a uint, then assigns to the StreamLabel as before. The problematic trace was only manifesting an issue at StreamReaderWriter.cs:153, but I searched for all locations where an int was being assigned to a StreamLabel in case any of them were following the same convention of using a signed type to represent unsigned data. SegmentedMemoryStreamReader.cs:135 and graph.cs:579 strongly appear to follow that pattern, I think it is unlikely in the other cases but the uint cast is a no-op in that case and does no harm.

The issue wasn't caught in testing because we don't have any ETW tests with sample inputs larger than 2GB. I added a unit test for parsing StreamLabel with a test case larger than 2GB to prevent regressions.