Fix IndexOfSetBit for the case of a negative pow2 input parameter - before the fix, the while loop would not execute because pow2 was negative, and thus the return value was 0.
The fix is to remove the test for pow2 > 0 from the while condition, but test for the special case pow2 == 0 before going into the loop, otherwise an infinite loop would result.

The call sites never call this routine with a 0 parameter, but the routine should be robust to a 0 input parameter nonetheless.