Add a new option in the right-click menu of the TreeView to open any TreeViewItems that are HTML reports in the machine's external browser.
To do this I had to fix a second issue - when you right-click an item in the tree view it is not automatically selected before you choose an item in the right-click menu. This meant that you could not reliably use TreeView.SelectedItem in the callback of the command that was chosen in the right-click menu. I fixed this by hooking in to the right-button-down event and navigating the visual tree to find the TreeViewItem related to the visual that was clicked.