Fix issue with IsKernelAddress on none Windows platforms.
https://github.com/microsoft/perfview/issues/1445 shows issues with
current schema to detect if a address is in kernel address space. On
none Windows platforms the implemneted rules are not correct causing
issues to display stack traces on nettrace files captured on some
configurations of none Windows platforms (Android is one example).

EventPipe nettrace files only include stack traces with managed frames
and only user mode code addresses, so addresses should never be in
kernel address space and since the heuristics missfire on none Windows
platforms, this commit adds logic to always detect addresses using a
process hooked up to an EventPipeEventSource to not be in kernel address
mode. None EventPipeEventSource event sources will continue to work as
before.