Change Download Link on Downloading.md
Getting the right release link to show up on Downloading.md involves creating the releases in the right order.  Try as I might, I still mess this up occasionally.  To address this, I have changed the link to just point to the releases page, so that folks can find the latest release directly.