Opt-in support for TaskID lookup directly from the event stream. (#1977)
By default, TaskID is not looked up from the event stream for each event,
and is instead pulled from event metadata or hardcoded in the parser.
This change allows users to ask ETWTraceEventSource to lookup the TaskID
from the event stream (present in the header) for each event.  This
enables tools that vary the TaskID for a given event ID to function
properly.