Fix surfacing of RelatedActivityID in real-time session dispatch. (#1980)
RelatedActivityID is stored in a GrowableArray as part of the TraceLog.
When the event is pushed in to the realTimeQueue in order to dispatch
it, the event is cloned.  Cloning creates a new instance of TraceEvent, and
allocates buffers that are owned by the new clone.  When the event is
dispatched, we are missing an assignment of myBuffer, which contains the
RelatedActivityID. Assign myBuffer prior to dispatch so that the RelatedActivityID
is correct.