Trim whitespace when parsing additional providers
Without this change, a very simple mistake such as an extra space in the
comma-separated list of providers will silently cause the provider to
not be enabled.  This is a very easy mistake to make if the list of
providers you want to enable it large and you are copying and pasting
the list from somewhere else and it gets word wrapped, etc.